+++ src/bika/lims/api/snapshot.py
@@ -144,9 +144,11 @@ def get_snapshot_version(obj, snapshot):
 
     :param obj: Content object
     :param snapshot: Snapshot dictionary
-    :returns: Index where the object is lcated
+    :returns: Index where the object is located
     """
     snapshots = get_snapshots(obj)
+    if snapshot not in snapshots:
+        return -1
     return snapshots.index(snapshot)